backend/requirements.txt
requests
psycopg2-binary
opensearch-py
python-dotenv
